Updated Identity Service & Created Gateway Service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Constracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -51,12 +52,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -72,6 +74,7 @@ namespace AuctionService.Controllers
                 new {auction.Id}, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id,UpdateAuctionDto updateAuctionDto)
         {
@@ -80,6 +83,8 @@ namespace AuctionService.Controllers
 
             if(auction == null) return NotFound();
 
+            if(auction.Seller != User.Identity.Name) return Forbid();
+            
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Color;
@@ -93,12 +98,9 @@ namespace AuctionService.Controllers
             if(result) return Ok();
 
             return BadRequest("Problem saving changes");
-
-
-            // TODO: check seller is equal to username
-
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -106,7 +108,7 @@ namespace AuctionService.Controllers
 
             if(auction == null) return NotFound();
 
-            // TODO check seller == username
+            if(auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
